+++ waitlist.go
@@ -30,7 +30,7 @@ type tWaitItem struct {
 	end  chan bool
 
 	protocol  int
-	port      int
+	port      uint64
 	timeoutUs int64
 	req       []uint8
 
@@ -45,7 +45,7 @@ var waitItemsMutex sync.Mutex
 // Call RPC同步调用
 // timeout是超时时间,单位:ms.为0表示不需要应答
 // 返回值是应答字节流和错误码.错误码非SystemOK表示调用失败
-func Call(protocol int, port int, dstIA uint64, rid int, timeout int, req []uint8) ([]uint8, ErrorCode) {
+func Call(protocol int, port uint64, dstIA uint64, rid int, timeout int, req []uint8) ([]uint8, ErrorCode) {
 	resp := CallAsync(protocol, port, dstIA, rid, timeout, req)
 	<-resp.Done
 	return resp.Bytes, resp.Error
@@ -54,7 +54,7 @@ func Call(protocol int, port int, dstIA uint64, rid int, timeout int, req []uint
 // CallAsync RPC异步调用
 // timeout是超时时间,单位:ms.为0表示不需要应答
 // 返回值中错误码非SystemOK表示调用失败
-func CallAsync(protocol int, port int, dstIA uint64, rid int, timeout int, req []uint8) *Resp {
+func CallAsync(protocol int, port uint64, dstIA uint64, rid int, timeout int, req []uint8) *Resp {
 	var resp Resp
 	resp.Done = make(chan *Resp, 10)
 
@@ -102,7 +102,7 @@ func CallAsync(protocol int, port int, dstIA uint64, rid int, timeout int, req [
 	return &resp
 }
 
-func waitlistSendFrame(protocol int, port int, dstIA uint64, code int, rid int, token int, data []uint8) {
+func waitlistSendFrame(protocol int, port uint64, dstIA uint64, code int, rid int, token int, data []uint8) {
 	if len(data) >= gSingleFrameSizeMax {
 		gBlockTx(protocol, port, dstIA, code, rid, token, data)
 		return
@@ -119,7 +119,7 @@ func waitlistSendFrame(protocol int, port int, dstIA uint64, code int, rid int, 
 }
 
 // gRxAckFrame 接收到ACK帧时处理函数
-func gRxAckFrame(protocol int, port int, srcIA uint64, frame *tFrame) {
+func gRxAckFrame(protocol int, port uint64, srcIA uint64, frame *tFrame) {
 	waitItemsMutex.Lock()
 	defer waitItemsMutex.Unlock()
 
@@ -137,7 +137,7 @@ func gRxAckFrame(protocol int, port int, srcIA uint64, frame *tFrame) {
 	}
 }
 
-func checkNodeAndDealAckFrame(protocol int, port int, srcIA uint64, frame *tFrame, node *list.Element) bool {
+func checkNodeAndDealAckFrame(protocol int, port uint64, srcIA uint64, frame *tFrame, node *list.Element) bool {
 	item := node.Value.(*tWaitItem)
 	if item.protocol != protocol || item.port != port || item.dstIA != srcIA || item.rid != frame.controlWord.rid ||
 		item.token != frame.controlWord.token {
@@ -152,7 +152,7 @@ func checkNodeAndDealAckFrame(protocol int, port int, srcIA uint64, frame *tFram
 }
 
 // gRxRstFrame 接收到RST帧时处理函数
-func gRxRstFrame(protocol int, port int, srcIA uint64, frame *tFrame) {
+func gRxRstFrame(protocol int, port uint64, srcIA uint64, frame *tFrame) {
 	waitItemsMutex.Lock()
 	defer waitItemsMutex.Unlock()
 
@@ -172,7 +172,7 @@ func gRxRstFrame(protocol int, port int, srcIA uint64, frame *tFrame) {
 
 // dealRstFrame 处理复位连接帧
 // 返回true表示节点符合条件
-func dealRstFrame(protocol int, port int, srcIA uint64, frame *tFrame, node *list.Element) bool {
+func dealRstFrame(protocol int, port uint64, srcIA uint64, frame *tFrame, node *list.Element) bool {
 	item := node.Value.(*tWaitItem)
 	if item.protocol != protocol || item.port != port || item.dstIA != srcIA || item.rid != frame.controlWord.rid ||
 		item.token != frame.controlWord.token {
